fix(wasm): engine_move_division u64→u32 — 修复 JS BigInt 转换错误
根因: WASM 函数 u64 参数, JS 调用时强制要求 BigInt(报错 Cannot convert 1 to BigInt)
修复: division_id 改 u32(师id不会超40亿), JS 直接传 Number 即可

+++ src/wasm_api.rs
@@ -84,8 +84,9 @@ pub unsafe extern "C" fn engine_set_province_controller(
 }
 
 /// 命令师移动到目标省(前端点选移动)
+/// 注: division_id 用 u32 而非 u64, 避免 JS 调用时 BigInt 转换问题
 #[no_mangle]
-pub extern "C" fn engine_move_division(division_id: u64, target: u32) {
+pub extern "C" fn engine_move_division(division_id: u32, target: u32) {
     ENGINE.with(|e| {
         let mut e = e.borrow_mut();
         let Engine { interp, world } = &mut *e;
